Use better semantic primitives for sprite loading.
* `u32` for pixel coordinates and dimensions.
* `f32` for offsets.

+++ examples/sprites_ordered/sprite.rs
@@ -4,17 +4,17 @@
 #[derive(Debug)]
 pub struct SpriteSheetDefinition {
     /// Width of each individual sprite on the sprite sheet.
-    pub sprite_w: f32,
+    pub sprite_w: u32,
     /// Height of each individual sprite on the sprite sheet.
-    pub sprite_h: f32,
+    pub sprite_h: u32,
     /// Number of rows in the sprite sheet.
     ///
     /// This is the number of sprites counting down the sheet.
-    pub row_count: usize,
+    pub row_count: u32,
     /// Number of columns in the sprite sheet.
     ///
     /// This is the number of sprites counting across the sheet.
-    pub column_count: usize,
+    pub column_count: u32,
     /// Whether or not there is a 1 pixel border between sprites.
     pub has_border: bool,
 }
@@ -30,10 +30,10 @@ impl SpriteSheetDefinition {
     /// * `column_count`: Number of columns in the sprite sheet.
     /// * `has_border`: Whether or not there is a 1 pixel border between sprites.
     pub fn new(
-        sprite_w: f32,
-        sprite_h: f32,
-        row_count: usize,
-        column_count: usize,
+        sprite_w: u32,
+        sprite_h: u32,
+        row_count: u32,
+        column_count: u32,
         has_border: bool,
     ) -> Self {
         SpriteSheetDefinition {
